# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My DOC API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/car-brands": {
            "get": {
                "tags": [
                    "Админские функции (Взаимодействие с брендом автомобиля)"
                ],
                "summary": "Список брэндов автомобилей",
                "operationId": "1f12b6867677daf48e35a038b3d5a4a6",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Бренд",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "created_from",
                        "in": "query",
                        "description": "Дата создания с",
                        "required": false,
                        "schema": {
                            "type": "date-time"
                        }
                    },
                    {
                        "name": "created_to",
                        "in": "query",
                        "description": "Дата создания до",
                        "required": false,
                        "schema": {
                            "type": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Porche"
                                                    },
                                                    "created_at": {
                                                        "type": "date-time",
                                                        "example": "2010-10-22"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Админские функции (Взаимодействие с брендом автомобиля)"
                ],
                "summary": "Создание названия бренда Автомобиля",
                "operationId": "547a7d5c2f8e5a1c1a32ee569dc9b278",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Porche"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Porche"
                                        },
                                        "created_at": {
                                            "type": "date-time",
                                            "example": "2020-10-22"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unautoruzed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/car-brands/{carBrand}": {
            "get": {
                "tags": [
                    "Админские функции (Взаимодействие с брендом автомобиля)"
                ],
                "summary": "Единичный брэнд",
                "operationId": "d05a8b67f7d4c791c761e7c2d60d0554",
                "parameters": [
                    {
                        "name": "carBrand",
                        "in": "path",
                        "description": "Id carBrand",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Porche"
                                        },
                                        "created_at": {
                                            "type": "date-time",
                                            "example": "2020-10-22"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Админские функции (Взаимодействие с брендом автомобиля)"
                ],
                "summary": "Удаление бренда",
                "operationId": "37c1bcc14ab6662b9ee1bec4581503b4",
                "parameters": [
                    {
                        "name": "carBrand",
                        "in": "path",
                        "description": "Id carBrand",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Админские функции (Взаимодействие с брендом автомобиля)"
                ],
                "summary": "Обновление названия бренда Автомобиля",
                "operationId": "ebbaded14991ebd7e2b51915221fb686",
                "parameters": [
                    {
                        "name": "carBrand",
                        "in": "path",
                        "description": "Car Brand ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Porche"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Porche"
                                        },
                                        "created_at": {
                                            "type": "date-time",
                                            "example": "2020-10-22"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unautoruzed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}